- FTS: reorganized pages packages - knxcomm: improved Memory* applications - knxcomm: implemented memory data connection adapter
git-svn-id: file:///usr/local/svn/freebus/trunk@2222 219b8a4f-0e80-dc11-b224-0019d16aef49

+++ freebus-fts-knxcomm/src/main/java/org/freebus/knxcomm/application/MemoryResponse.java
@@ -24,7 +24,7 @@ public class MemoryResponse extends MemoryData
     * @throws IllegalArgumentException if the supplied memory data has more than
     *            63 bytes.
     */
-   protected MemoryResponse(int address, int[] data)
+   protected MemoryResponse(int address, byte[] data)
    {
       super(address, data);
    }
@@ -38,7 +38,7 @@ public class MemoryResponse extends MemoryData
     * @throws IllegalArgumentException if the supplied memory data has more than
     *            63 bytes.
     */
-   protected MemoryResponse(MemoryLocation location, int[] data)
+   protected MemoryResponse(MemoryLocation location, byte[] data)
    {
       super(location, data);
    }
