+++ freebus-fts-knxcomm/src/test/java/org/freebus/knxcomm/application/TestMemoryWrite.java
@@ -26,17 +26,17 @@ public class TestMemoryWrite
    @Test
    public final void testMemoryWriteIntIntArray()
    {
-      final int data[] = new int[] { 1, 2, 3 };
+      final byte data[] = new byte[] { 1, 2, 3 };
       final MemoryWrite app = new MemoryWrite(1230, data);
 
       assertNotNull(app.toString());
 
       assertEquals(1230, app.getAddress());
       assertEquals(3, app.getCount());
-      assertArrayEquals(new int[] { 1, 2, 3 }, app.getData());
+      assertArrayEquals(new byte[] { 1, 2, 3 }, app.getData());
 
       data[0] = 0;
-      assertArrayEquals(new int[] { 1, 2, 3 }, app.getData());
+      assertArrayEquals(new byte[] { 1, 2, 3 }, app.getData());
 
       new MemoryWrite(10900, null);
    }
@@ -44,6 +44,6 @@ public class TestMemoryWrite
    @Test(expected = IllegalArgumentException.class)
    public final void testMemoryWriteIntIntArrayTooLarge()
    {
-      new MemoryWrite(1230, new int[64]);
+      new MemoryWrite(1230, new byte[64]);
    }
 }
